requirements.txt
# Dependencias principales - Compatibles con Python 3.13.2
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.2
openpyxl==3.1.5
plotly==5.24.1
requests==2.32.3
python-dotenv==1.0.1

# Para análisis de datos
scikit-learn==1.5.2
seaborn==0.13.2
matplotlib==3.9.2

# Para interacción con Ollama
httpx==0.27.2
aiohttp==3.10.10